Restructure package around new capabilities in Base
This is a major redesign:
- requires julia-0.7
- relies on new include callbacks and extra information stored in *.ji files to get the list of files to track
- splits code base into multiple files for easier navigation and understanding
- adopts the Base.root_module interface for working with modules
