+++ test/runtests.jl
@@ -1,5 +1,5 @@
 using Revise
-using Base.Test
+using Test
 using DataStructures: OrderedSet
 using Compat
 
@@ -15,7 +15,7 @@ to_remove = String[]
         exs
     end
 
-    @static if is_apple()
+    @static if Sys.isapple()
         yry() = (sleep(1.1); revise(); sleep(1.1))
     else
         yry() = (sleep(0.1); revise(); sleep(0.1))
@@ -46,7 +46,7 @@ g(x) = 2
 h{x) = 3  # error
 k(x) = 4
 """,
-                                            :test, 1, Main, tempdir())
+                                            :test, 1, Main)
                 @test convert(Revise.RelocatableExpr, :(g(x) = 2)) ∈ md[Main]
             end
         end
@@ -62,8 +62,8 @@ k(x) = 4
         # test the "mistakes"
         @test ReviseTest.cube(2) == 16
         @test ReviseTest.Internal.mult3(2) == 8
-        oldmd = Revise.parse_source(fl1, Main, dirname(fl1))
-        newmd = Revise.parse_source(fl2, Main, nothing)
+        oldmd = Revise.parse_source(fl1, Main).second
+        newmd = Revise.parse_source(fl2, Main).second
         revmd = Revise.revised_statements(newmd, oldmd)
         @test length(revmd) == 2
         @test haskey(revmd, ReviseTest) && haskey(revmd, ReviseTest.Internal)
@@ -84,7 +84,7 @@ k(x) = 4
         @test ReviseTest.Internal.mult3(2) == 6
 
         # Backtraces
-        newmd = Revise.parse_source(fl3, Main, nothing)
+        newmd = Revise.parse_source(fl3, Main).second
         revmd = Revise.revised_statements(newmd, oldmd)
         Revise.eval_revised(revmd)
         try
@@ -569,8 +569,9 @@ revise_f(x) = 2
         cd(curdir)
 
         # Tracking Base
-        Revise.track(Base)
-        @test any(k->endswith(k, "number.jl"), keys(Revise.file2modules))
+        # FIXME
+        # Revise.track(Base)
+        # @test any(k->endswith(k, "number.jl"), keys(Revise.file2modules))
     end
 
     @testset "Cleanup" begin
@@ -586,7 +587,7 @@ revise_f(x) = 2
                 yry()
             end
         end
-        if !is_apple()
+        if !Sys.isapple()
             @test contains(read(warnfile, String), "is not an existing directory")
         end
         rm(warnfile)
